спринта №10 версия 1
