+++ app/src/main/java/com/example/playlistmaker/SearchActivity.kt
@@ -4,18 +4,12 @@ import android.content.Context
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
 import android.widget.EditText
 import android.widget.ImageView
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.bumptech.glide.load.resource.bitmap.TransformationUtils.centerCrop
 
 class SearchActivity : AppCompatActivity() {
     companion object {
@@ -27,9 +21,9 @@ class SearchActivity : AppCompatActivity() {
     private val tracks = TrackAdapter(
         listOf(
             Track(
-                "Smells Like Teen Spirittttttttttttttttttttttttttttttttttttttttt",
-                "Nirvanaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "5:0111111111111111111111111111111111111111111",
+                "Smells Like Teen Spirit",
+                "Nirvana",
+                "5:01",
                 "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
             ),
 
@@ -53,33 +47,6 @@ class SearchActivity : AppCompatActivity() {
                 "5:33",
                 "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
             ),
-
-            Track(
-                "Sweet Child O'Mine",
-                "Guns N' Roses",
-                "5:03",
-                "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
-            ),
-
-            Track(
-                "Stayin' Alive",
-                "Bee Gees",
-                "4:10",
-                "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
-            ),
-
-            Track(
-                "Whole Lotta Love",
-                "Led Zeppelin",
-                "5:33",
-                "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
-            ),
-
-            Track(
-                "Sweet Child O'Mine",
-                "Guns N' Roses",
-                "5:03",
-                "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg")
         )
     )
 
@@ -156,45 +123,3 @@ fun View.hideKeyboard() {
     )
 }
 
-class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    private val albumImage: ImageView = itemView.findViewById(R.id.album)
-    private val songText: TextView = itemView.findViewById(R.id.song_name)
-    private val authorText: TextView = itemView.findViewById(R.id.author)
-    private val timeText: TextView = itemView.findViewById(R.id.time)
-
-    fun bind(model: Track) {
-        Glide.with(itemView)
-            .load(model.artworkUrl100)
-            .centerCrop()
-            .placeholder(R.drawable.placeholder)
-            .transform(RoundedCorners(2))
-            .into(albumImage)
-        songText.text = model.trackName
-        authorText.text = model.artistName
-        timeText.text = model.trackTime
-
-    }
-
-}
-
-class TrackAdapter(
-    private val data: List<Track>
-) : RecyclerView.Adapter<TrackViewHolder>() {
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int
-    ): TrackViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.items_view, parent, false)
-        return TrackViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
-        holder.bind(data[position])
-    }
-
-    override fun getItemCount(): Int {
-        return data.size
-    }
-}
