+++ app/src/main/res/layout/items_view.xml
@@ -16,9 +16,12 @@
         android:scaleType="fitXY" />
 
     <LinearLayout
+        android:id="@+id/item1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_weight="1"
         android:orientation="vertical">
+
 
         <TextView
             android:id="@+id/song_name"
@@ -26,37 +29,40 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="8dp"
             android:layout_marginTop="14dp"
+            android:ellipsize="end"
             android:fontFamily="@font/ys_display_regular"
             android:gravity="center_vertical"
+            android:maxLines="1"
             android:textColor="@color/black"
             android:textSize="16sp"
             tools:text="Yesterday (Remastered 2009)" />
 
         <LinearLayout
+            android:id="@+id/item2"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginStart="8dp"
             android:orientation="horizontal">
 
             <TextView
                 android:id="@+id/author"
-                android:layout_width="wrap_content"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="14dp"
-                android:ellipsize="middle"
+                android:layout_gravity="center_horizontal"
+                android:layout_marginTop="1dp"
+                android:layout_weight="1"
+                android:ellipsize="end"
                 android:fontFamily="@font/ys_display_regular"
-                android:gravity="center_vertical"
-                android:lines="1"
+                android:maxLines="1"
                 android:textColor="@color/text_author"
                 android:textSize="16sp"
                 tools:text="The Beatles" />
 
             <ImageView
-                android:id="@+id/back_button"
+                android:id="@+id/dot"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_gravity="bottom"
-                android:layout_marginBottom="5dp"
+                android:layout_gravity="center_vertical"
                 android:src="@drawable/dot"
                 tools:ignore="ContentDescription" />
 
@@ -64,8 +70,7 @@
                 android:id="@+id/time"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="14dp"
-                android:ellipsize="middle"
+                android:layout_weight="0"
                 android:fontFamily="@font/ys_display_regular"
                 android:gravity="center_vertical"
                 android:lines="1"
@@ -74,7 +79,9 @@
                 tools:text="2:55" />
 
         </LinearLayout>
+
     </LinearLayout>
+
 
     <ImageView
         android:id="@+id/arrow_forward"
@@ -84,6 +91,7 @@
         android:layout_marginTop="18dp"
         android:layout_marginEnd="12dp"
         android:layout_marginBottom="19dp"
+        android:layout_weight="0"
         android:src="@drawable/arrow_forward"
         tools:ignore="ContentDescription" />
 </LinearLayout>
